Add Compiled associated data family to RegexBackend
Also add a corresponding: `match :: Compiled s x -> s -> Maybe x` method.

This change abstracts away how matching is actually performed. Currently, the "compiled" version for ByteStrings
and Strings is still a function (wrapped in a newtype). But it will allow me to seamlessly (API-wise) replace it
with an optimized data structure.